refactor to hibp controller and test
