+++ tests/scan.test.js
@@ -4,25 +4,11 @@ const sha1 = require("../sha1-utils");
 const HIBP = require("../hibp");
 const scan = require("../controllers/scan");
 
+const { testBreaches } = require ("./test-breaches");
 require("./resetDB");
 
 
 jest.mock("../hibp");
-
-const testBreaches = [
-  {
-    Title: "Test",
-    Name: "Test",
-    Domain: "test.com",
-    BreachDate: "2012-12-21",
-  },
-  {
-    Title: "Test2",
-    Name: "Test2",
-    Domain: "test2.com",
-    BreachDate: "2016-11-08",
-  },
-];
 
 
 test("scan GET redirects to home", () => {
