+++ tests/test-breaches.js
@@ -1,0 +1,20 @@
+"use strict";
+
+const testBreaches = [
+  {
+    Title: "Test",
+    Name: "Test",
+    Domain: "test.com",
+    BreachDate: "2012-12-21",
+  },
+  {
+    Title: "Test2",
+    Name: "Test2",
+    Domain: "test2.com",
+    BreachDate: "2016-11-08",
+  },
+];
+
+module.exports = {
+  testBreaches,
+};
